+++ scripts/accept_visual_regression.js
@@ -6,8 +6,8 @@ const fs = require("fs/promises");
 const { copy } = require("../src/lib/utils/fs");
 const { join } = require("path");
 
-const expectedDir = join(__dirname, "../dist/tmp/.reg/expected");
-const outputDir = join(__dirname, "../dist/tmp/__screenshots__");
+const expectedDir = join(__dirname, "../tmp/baseline");
+const outputDir = join(__dirname, "../tmp/screenshots");
 
 fs.rm(expectedDir, { recursive: true, force: true })
     .then(() => copy(outputDir, expectedDir))