+++ src/test/converter2/issues/gh2087.ts
@@ -1,0 +1,9 @@
+/** Foo type comment */
+export type Foo = () => number;
+
+export class Bar {
+    /**
+     * {@inheritDoc Foo:type}
+     */
+    x = 1;
+}